Revert "Enable CONFIG_NARROWBAND_SUPPORTING & disable it by default in moduleparam"
It seems that the module param controlled by this macro can not work with 10M firmware. The card failed to init.
I guess that they had failed (or not finish yet) on setting MAC freq in driver, so they used a modified firmware instead.
However, a spectrum analyzer is needed to see what's going on when card working with rtw_nb_config=0 & 10M firmware,
and unfortunately I don't have any at the time...

This reverts commit abdb58b198ec77fcb138dca12e13953aca0e27f4.